// src/api-toggl/types.gen.ts
export type ModelsTimeEntry = {
    billable?: boolean;
    client_id?: number;
    client_name?: string;
    created_at?: string;
    created_with?: string;
    deleted_at?: string;
    description?: string;
    duration?: number;
    id?: number;
    project_id?: number;
    project_name?: string;
    start?: string;
    tags?: Array<string>;
    task_id?: number;
    task_name?: string;
    workspace_id?: number;
    workspace_name?: string;
};