// src/api-redmine/types.gen.ts
export type IdName = {
    id: number;
    name: string;
};

export type IssueStatus = {
    id: number;
    name: string;
    is_closed: boolean;
};

export type CustomFieldValue = {
    id: number;
    name: string;
    multiple?: boolean;
    [key: string]: unknown | number | string | boolean | undefined;
};
// ...
export type TimeEntry = {
    id: number;
    project?: IdName;
    issue?: {
        id: number;
    };
    user: IdName;
    activity: IdName;
    hours: number;
    comments: string | null;
    spent_on: string;
    created_on: string;
    updated_on: string;
    custom_fields?: Array<CustomFieldValue>;
};
// ...
export type Search = {
    id: number;
    title: string;
    type: string;
    url: string;
    description: string;
    datetime: string;
};